# src/api/api模板文件/full_openapi.yaml
openapi: 3.1.0
info:
  title: 待办事项后端
  description: 一个简洁的 Todo-List API，用于演示 FastAPI + SQLAlchemy 的中文化接口。
  version: 0.1.0
paths:
...
  /pomodoros/:
    get:
      operationId: list_pomodoros_pomodoros__get
      summary: 番茄钟列表
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PomodoroOut'
                title: Response List Pomodoros Pomodoros  Get
      tags:
        - 番茄钟
    post:
      operationId: start_pomodoro_pomodoros__post
      summary: 开始番茄钟
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PomodoroStart'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PomodoroOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      tags:
        - 番茄钟
  '/pomodoros/{session_id}/stop':
    patch:
      operationId: stop_pomodoro_pomodoros__session_id__stop_patch
      summary: 结束番茄钟
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: integer
            title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PomodoroOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      tags:
        - 番茄钟
components:
  schemas:
    HTTPValidationError:
      type: object
      properties:
        detail:
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
      title: HTTPValidationError
...
    PomodoroOut:
      description: 番茄钟详情
      type: object
      properties:
        id:
          type: integer
          title: Id
        start_at:
          type: string
          format: date-time
          title: Start At
        end_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
          title: End At
        work_minutes:
          type: integer
          title: Work Minutes
        break_minutes:
          type: integer
          title: Break Minutes
        planned_seconds:
          type: integer
          title: Planned Seconds
        actual_seconds:
          anyOf:
            - type: integer
            - type: 'null'
          title: Actual Seconds
        completed:
          type: boolean
          title: Completed
      required:
        - id
        - start_at
        - end_at
        - work_minutes
        - break_minutes
        - planned_seconds
        - actual_seconds
        - completed
      title: 番茄钟详情
    PomodoroStart:
      description: 番茄钟启动
      type: object
      properties:
        work_minutes:
          type: integer
          default: 25
          maximum: 120
          minimum: 5
          title: 专注时长（分钟）
        break_minutes:
          type: integer
          default: 5
          maximum: 60
          minimum: 1
          title: 休息时长（分钟）
      title: 番茄钟启动
...
    ValidationError:
      type: object
      properties:
        loc:
          type: array
          items:
            anyOf:
              - type: string
              - type: integer
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      required:
        - loc
        - msg
        - type
      title: ValidationError
...